+++ sklearn/cluster/_feature_agglomeration.py
@@ -19,7 +19,9 @@ class AgglomerationTransform(TransformerMixin):
     A class for feature agglomeration via the transform interface
     """
 
-    def transform(self, X, pooling_func=np.mean):
+    pooling_func = np.mean
+
+    def transform(self, X, pooling_func=None):
         """
         Transform a new matrix using the built clustering
 
@@ -39,6 +41,11 @@ class AgglomerationTransform(TransformerMixin):
         Y : array, shape = [n_samples, n_clusters] or [n_clusters]
             The pooled values for each feature cluster.
         """
+        if pooling_func is not None:
+            warnings.warn("The pooling_func parameter is deprecated since 0.15 and will be "
+                "removed in 0.18. Pass it to the constructor instead.", DeprecationWarning)
+        else:
+            pooling_func = self.pooling_func
         X = array2d(X)
         nX = []
         if len(self.labels_) != X.shape[1]: