+++ sklearn/externals/joblib/memory.py
@@ -406,10 +406,8 @@ class MemorizedFunc(Logger):
         # concurrent joblibs removing the file or the directory
         try:
             mkdirp(output_dir)
-            json.dump(
-                input_repr,
-                file(os.path.join(output_dir, 'input_args.json'), 'w'),
-                )
+            with open(os.path.join(output_dir, 'input_args.json'), 'w') as f:
+                json.dump(input_repr, f)
         except:
             pass
         return input_repr