+++ sklearn/feature_extraction/text.py
@@ -903,8 +903,11 @@ class TfidfTransformer(BaseEstimator, TransformerMixin):
     informative than features that occur in a small fraction of the training
     corpus.
 
-    In the SMART notation used in IR, this class implements several tf-idf
-    variants:
+    The actual formula used for tf-idf is tf * (idf + 1) = tf + tf * idf,
+    instead of tf * idf. The effect of this is that terms with zero idf, i.e.
+    that occur in all documents of a training set, will not be entirely
+    ignored. The formulas used to compute tf and idf depend on parameter
+    settings that correspond to the SMART notation used in IR, as follows:
 
     Tf is "n" (natural) by default, "l" (logarithmic) when sublinear_tf=True.
     Idf is "t" when use_idf is given, "n" (none) otherwise.
@@ -962,7 +965,8 @@ class TfidfTransformer(BaseEstimator, TransformerMixin):
             df += int(self.smooth_idf)
             n_samples += int(self.smooth_idf)
 
-            # avoid division by zeros for features that occur in all documents
+            # log1p instead of log makes sure terms with zero idf don't get
+            # suppressed entirely
             idf = np.log(float(n_samples) / df) + 1.0
             self._idf_diag = sp.spdiags(idf,
                                         diags=0, m=n_features, n=n_features)
@@ -1120,9 +1124,9 @@ class TfidfVectorizer(CountVectorizer):
         given, a vocabulary is determined from the input documents.
 
     binary : boolean, False by default.
-        If True, all non zero counts are set to 1. This is useful for discrete
-        probabilistic models that model binary events rather than integer
-        counts.
+        If True, all non-zero term counts are set to 1. This does not mean
+        outputs will have only 0/1 values, only that the tf term in tf-idf
+        is binary. (Set idf and normalization to False to get 0/1 outputs.)
 
     dtype : type, optional
         Type of the matrix returned by fit_transform() or transform().
@@ -1170,7 +1174,7 @@ class TfidfVectorizer(CountVectorizer):
             preprocessor=preprocessor, tokenizer=tokenizer, analyzer=analyzer,
             stop_words=stop_words, token_pattern=token_pattern,
             ngram_range=ngram_range, max_df=max_df, min_df=min_df,
-            max_features=max_features, vocabulary=vocabulary, binary=False,
+            max_features=max_features, vocabulary=vocabulary, binary=binary,
             dtype=dtype)
 
         self._tfidf = TfidfTransformer(norm=norm, use_idf=use_idf,
@@ -1249,4 +1253,4 @@ class TfidfVectorizer(CountVectorizer):
         vectors : sparse matrix, [n_samples, n_features]
         """
         X = super(TfidfVectorizer, self).transform(raw_documents)
-        return self._tfidf.transform(X, copy)
+        return self._tfidf.transform(X, copy=False)