+++ sklearn/tree/_utils.pyx
@@ -4,6 +4,7 @@
 
 # Authors: Gilles Louppe <g.louppe@gmail.com>
 #          Peter Prettenhofer <peter.prettenhofer@gmail.com>
+#          Arnaud Joly <arnaud.v.joly@gmail.com>
 #
 # Licence: BSD 3 clause
 
@@ -22,6 +23,7 @@ cdef inline void copy_stack(StackRecord* a, StackRecord* b) nogil:
     a.parent = b.parent
     a.is_left = b.is_left
     a.impurity = b.impurity
+    a.n_constant_features = b.n_constant_features
 
 
 cdef class Stack:
@@ -54,7 +56,8 @@ cdef class Stack:
         return self.top <= 0
 
     cdef int push(self, SIZE_t start, SIZE_t end, SIZE_t depth, SIZE_t parent,
-                   bint is_left, double impurity) nogil:
+                  bint is_left, double impurity,
+                  SIZE_t n_constant_features) nogil:
         """Push a new element onto the stack.
 
         Returns 0 if successful; -1 on out of memory error.
@@ -79,6 +82,7 @@ cdef class Stack:
         stack[top].parent = parent
         stack[top].is_left = is_left
         stack[top].impurity = impurity
+        stack[top].n_constant_features = n_constant_features
 
         # Increment stack pointer
         self.top = top + 1
@@ -115,6 +119,8 @@ cdef inline void copy_heap(PriorityHeapRecord* a, PriorityHeapRecord* b) nogil:
     a.depth = b.depth
     a.is_leaf = b.is_leaf
     a.impurity = b.impurity
+    a.impurity_left = b.impurity_left
+    a.impurity_right = b.impurity_right
     a.improvement = b.improvement
 
 
@@ -127,7 +133,8 @@ cdef void swap_heap(PriorityHeapRecord* heap, SIZE_t a, SIZE_t b) nogil:
 
 
 cdef void heapify_up(PriorityHeapRecord* heap, SIZE_t pos) nogil:
-    """Restore heap invariant parent.improvement > child.improvement from ``pos`` upwards. """
+    """Restore heap invariant parent.improvement > child.improvement from
+       ``pos`` upwards. """
     if pos == 0:
         return
 
@@ -138,15 +145,20 @@ cdef void heapify_up(PriorityHeapRecord* heap, SIZE_t pos) nogil:
         heapify_up(heap, parent_pos)
 
 
-cdef void heapify_down(PriorityHeapRecord* heap, SIZE_t pos, SIZE_t heap_length) nogil:
-    """Restore heap invariant parent.improvement > children.improvement from ``pos`` downwards. """
+cdef void heapify_down(PriorityHeapRecord* heap, SIZE_t pos,
+                       SIZE_t heap_length) nogil:
+    """Restore heap invariant parent.improvement > children.improvement from
+       ``pos`` downwards. """
     cdef SIZE_t left_pos = 2 * (pos + 1) - 1
     cdef SIZE_t right_pos = 2 * (pos + 1)
     cdef SIZE_t largest = pos
 
-    if left_pos < heap_length and heap[left_pos].improvement > heap[largest].improvement:
+    if (left_pos < heap_length and
+            heap[left_pos].improvement > heap[largest].improvement):
         largest = left_pos
-    if right_pos < heap_length and heap[right_pos].improvement > heap[largest].improvement:
+
+    if (right_pos < heap_length and
+            heap[right_pos].improvement > heap[largest].improvement):
         largest = right_pos
 
     if largest != pos:
@@ -188,8 +200,9 @@ cdef class PriorityHeap:
         return self.heap_ptr <= 0
 
     cdef int push(self, SIZE_t node_id, SIZE_t start, SIZE_t end, SIZE_t pos,
-                   SIZE_t depth, bint is_leaf, double improvement,
-                   double impurity) nogil:
+                  SIZE_t depth, bint is_leaf, double improvement,
+                  double impurity, double impurity_left,
+                  double impurity_right) nogil:
         """Push record on the priority heap.
 
         Returns 0 if successful; -1 on out of memory error.
@@ -217,6 +230,8 @@ cdef class PriorityHeap:
         heap[heap_ptr].depth = depth
         heap[heap_ptr].is_leaf = is_leaf
         heap[heap_ptr].impurity = impurity
+        heap[heap_ptr].impurity_left = impurity_left
+        heap[heap_ptr].impurity_right = impurity_right
         heap[heap_ptr].improvement = improvement
 
         # Heapify up