+++ sklearn/utils/tests/test_extmath.py
@@ -20,6 +20,7 @@ from sklearn.utils.testing import assert_raises
 
 from sklearn.utils.extmath import density
 from sklearn.utils.extmath import logsumexp
+from sklearn.utils.extmath import norm, squared_norm
 from sklearn.utils.extmath import randomized_svd
 from sklearn.utils.extmath import row_norms
 from sklearn.utils.extmath import weighted_mode
@@ -123,6 +124,16 @@ def test_randomized_svd_low_rank():
     # compute the singular values of X using the fast approximate method
     Ua, sa, Va = randomized_svd(X, k)
     assert_almost_equal(s[:rank], sa[:rank])
+
+
+def test_norm_squared_norm():
+    X = np.random.RandomState(42).randn(50, 63)
+    X *= 100        # check stability
+    X += 200
+
+    assert_almost_equal(np.linalg.norm(X.ravel()), norm(X))
+    assert_almost_equal(norm(X) ** 2, squared_norm(X), decimal=6)
+    assert_almost_equal(np.linalg.norm(X), np.sqrt(squared_norm(X)), decimal=6)
 
 
 def test_row_norms():
@@ -289,6 +300,9 @@ def test_logistic_sigmoid():
 
 def test_fast_dot():
     """Check fast dot blas wrapper function"""
+    if fast_dot is np.dot:
+        return
+
     rng = np.random.RandomState(42)
     A = rng.random_sample([2, 10])
     B = rng.random_sample([2, 10])
@@ -302,28 +316,28 @@ def test_fast_dot():
     if has_blas:
         # Test _fast_dot for invalid input.
 
-            # Maltyped data.
-            for dt1, dt2 in [['f8', 'f4'], ['i4', 'i4']]:
-                assert_raises(ValueError, _fast_dot, A.astype(dt1),
-                              B.astype(dt2).T)
+        # Maltyped data.
+        for dt1, dt2 in [['f8', 'f4'], ['i4', 'i4']]:
+            assert_raises(ValueError, _fast_dot, A.astype(dt1),
+                          B.astype(dt2).T)
 
-            # Malformed data.
+        # Malformed data.
 
-            ## ndim == 0
-            E = np.empty(0)
-            assert_raises(ValueError, _fast_dot, E, E)
+        ## ndim == 0
+        E = np.empty(0)
+        assert_raises(ValueError, _fast_dot, E, E)
 
-            ## ndim == 1
-            assert_raises(ValueError, _fast_dot, A, A[0])
+        ## ndim == 1
+        assert_raises(ValueError, _fast_dot, A, A[0])
 
-            ## ndim > 2
-            assert_raises(ValueError, _fast_dot, A.T, np.array([A, A]))
+        ## ndim > 2
+        assert_raises(ValueError, _fast_dot, A.T, np.array([A, A]))
 
-            ## min(shape) == 1
-            assert_raises(ValueError, _fast_dot, A, A[0, :][None, :])
+        ## min(shape) == 1
+        assert_raises(ValueError, _fast_dot, A, A[0, :][None, :])
 
-            # test for matrix mismatch error
-            assert_raises(ValueError, _fast_dot, A, A)
+        # test for matrix mismatch error
+        assert_raises(ValueError, _fast_dot, A, A)
 
     # Test cov-like use case + dtypes.
     for dtype in ['f8', 'f4']:
@@ -333,15 +347,15 @@ def test_fast_dot():
         #  col < row
         C = np.dot(A.T, A)
         C_ = fast_dot(A.T, A)
-        assert_almost_equal(C, C_)
+        assert_almost_equal(C, C_, decimal=5)
 
         C = np.dot(A.T, B)
         C_ = fast_dot(A.T, B)
-        assert_almost_equal(C, C_)
+        assert_almost_equal(C, C_, decimal=5)
 
         C = np.dot(A, B.T)
         C_ = fast_dot(A, B.T)
-        assert_almost_equal(C, C_)
+        assert_almost_equal(C, C_, decimal=5)
 
     # Test square matrix * rectangular use case.
     A = rng.random_sample([2, 2])
@@ -351,11 +365,11 @@ def test_fast_dot():
 
         C = np.dot(A, B)
         C_ = fast_dot(A, B)
-        assert_almost_equal(C, C_)
+        assert_almost_equal(C, C_, decimal=5)
 
         C = np.dot(A.T, B)
         C_ = fast_dot(A.T, B)
-        assert_almost_equal(C, C_)
+        assert_almost_equal(C, C_, decimal=5)
 
     if has_blas:
         for x in [np.array([[d] * 10] * 2) for d in [np.inf, np.nan]]: